feat: 불필요한 커스텀 훅 삭제 및 베인페이지 리팩토링
