+++ src/App.js
@@ -1,9 +1,13 @@
 import styled from "styled-components";
-import React from 'react';
-import CopySolution from "./pages/CopySolution";
+import React from "react";
+import CopySolution from "./pages/Home";
 
 function App() {
-    return <Wrapper><CopySolution/></Wrapper>
+  return (
+    <Wrapper>
+      <CopySolution />
+    </Wrapper>
+  );
 }
 
 export default App;
