+++ src/pages/Home/components/index.js
@@ -1,3 +1,3 @@
-export { default as Retry } from "./Retry";
+export { default as Retry } from "./Error";
 export { default as Success } from "./Success";
 export { default as NoSolution } from "./NoSolution";
