+++ src/shared/utils/chrome.js
@@ -1,0 +1,5 @@
+export const requestToContentScript = async (key) => {
+  const queryOptions = { active: true, lastFocusedWindow: true };
+  const [currentTab] = await chrome.tabs.query(queryOptions);
+  return await chrome.tabs.sendMessage(currentTab.id, key);
+};
